Split Module into header and body files
